attempt to write edges to vtk for debugging
The goal was to see if meshes generated with mesh_top1 had classification preserved via the ids for entities.

From what I can tell the edges bounding triangles are not explicitly
stored as entities in the structure.  'mesh.edge2' stores edges
classified on boundaries and 'mesh.tria3' stores indices of vertices
bounding each triangle.  Assuming the ids of those entities are correct
we should be OK.

The .msh files appear to contain the same data that is stored in the
'mesh' object described above.